requirements-prod.txt
# Production requirements for p4td_backend
# Install with: pip install -r requirements-prod.txt

# Django core
Django>=5.0,<6.0
djangorestframework>=3.14
djoser>=2.2
django-cors-headers>=4.3

# Production server
gunicorn>=21.0

# Database
psycopg2-binary>=2.9

# AWS S3 storage
django-storages[s3]>=1.14
boto3>=1.34

# Static files
whitenoise>=6.6

# Environment variables
python-dotenv>=1.0

# Image processing (for thumbnails)
Pillow>=10.0
